Admin Panel and Frontend fixes
Fixes in the admin panel. Able to create users, categories and upload
images from the index view.
Deleted unused views for users, media.
Created preview view for posts and added updated and created fields to
the preview. Need to be more cute.
Fixed the issue with the css in the admin sections. There were “:”
after classes on style.css.
Created categories admin panel.
Put everything that works for posts under the posts menu.